# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: FuelGo Islamabad - Fuel Delivery Platform API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /pumps:
    get:
      operationId: listPumps
      tags: [pumps]
      summary: List all approved petrol pumps
      parameters:
        - name: lat
          in: query
          schema:
            type: number
        - name: lng
          in: query
          schema:
            type: number
        - name: radius
          in: query
          schema:
            type: number
        - name: search
          in: query
          schema:
            type: string
        - name: fuelType
          in: query
          schema:
            type: string
      responses:
        "200":
          description: Pumps list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/PetrolPump"
  /pumps/{id}:
    get:
      operationId: getPump
      tags: [pumps]
      summary: Get petrol pump details
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Pump details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PetrolPump"
        "404":
          description: Pump not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /pumps/my:
    get:
      operationId: getMyPump
      tags: [pumps]
      summary: Get authenticated pump owner's pump profile
      responses:
        "200":
          description: Own pump details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PetrolPump"
        "404":
          description: No pump found for this owner
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

    patch:
      operationId: updateMyPump
      tags: [pumps]
      summary: Update authenticated pump owner's pump profile
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PumpProfileUpdate"
      responses:
        "200":
          description: Pump updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PetrolPump"
...
  /pumps/{pumpId}/reviews:
    get:
      operationId: getPumpReviews
      tags: [reviews]
      summary: Get reviews for a pump
      parameters:
        - name: pumpId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Reviews list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Review"

    post:
      operationId: createReview
      tags: [reviews]
      summary: Submit a review for a pump
      parameters:
        - name: pumpId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ReviewInput"
      responses:
        "201":
          description: Review submitted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Review"
        "400":
          description: Invalid review or already reviewed
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  # ─── ADMIN ──────────────────────────────────────────────────────────────────
...
components:
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
      required: [error]
...
    PetrolPump:
      type: object
      required: [id, businessName, ownerName, phone, address, status, createdAt]
      properties:
        id:
          type: integer
        ownerId:
          type: integer
        businessName:
          type: string
        ownerName:
          type: string
        phone:
          type: string
        email:
          type: ["string", "null"]
        address:
          type: string
        latitude:
          type: ["number", "null"]
        longitude:
          type: ["number", "null"]
        operatingHoursStart:
          type: ["string", "null"]
        operatingHoursEnd:
          type: ["string", "null"]
        deliveryRadius:
          type: ["number", "null"]
        deliveryCharges:
          type: ["number", "null"]
        description:
          type: ["string", "null"]
        logoUrl:
          type: ["string", "null"]
        status:
          type: string
          enum: [pending, approved, rejected, suspended]
        rejectionReason:
          type: ["string", "null"]
        averageRating:
          type: ["number", "null"]
        totalReviews:
          type: ["integer", "null"]
        isOpen:
          type: ["boolean", "null"]
        fuelPrices:
          type: array
          items:
            $ref: "#/components/schemas/FuelPrice"
        createdAt:
          type: string
    PumpProfileUpdate:
      type: object
      properties:
        businessName:
          type: string
        description:
          type: string
        address:
          type: string
        latitude:
          type: number
        longitude:
          type: number
        operatingHoursStart:
          type: string
        operatingHoursEnd:
          type: string
        deliveryRadius:
          type: number
        deliveryCharges:
          type: number
        logoUrl:
          type: string
        isOpen:
          type: boolean
...
    FuelPrice:
      type: object
      required: [id, pumpId, fuelType, pricePerLiter, isAvailable]
      properties:
        id:
          type: integer
        pumpId:
          type: integer
        fuelType:
          type: string
          enum: [petrol, diesel, hi_octane]
        pricePerLiter:
          type: number
        isAvailable:
          type: boolean
        updatedAt:
          type: string
...
    ReviewInput:
      type: object
      required: [rating, orderId]
      properties:
        orderId:
          type: integer
        rating:
          type: integer
          minimum: 1
          maximum: 5
        comment:
          type: ["string", "null"]
    Review:
      type: object
      required: [id, pumpId, customerId, rating, createdAt]
      properties:
        id:
          type: integer
        pumpId:
          type: integer
        customerId:
          type: integer
        orderId:
          type: integer
        rating:
          type: integer
        comment:
          type: ["string", "null"]
        customerName:
          type: ["string", "null"]
        createdAt:
          type: string

    # ── Admin ────────────────────────────────────────────────────────────────
...